Fixed Addition and Removal
I changed the functions of the addInventory button so that it now adds inventory as desired. Removal is a different story, I still have not implemented that yet. I also made the JList work on a index system, where every row is associated to an index, which is then associated with an item in the InventoryArrayList<Inventory>.