Removed Unnecessary Code
I got rid of the InventoryArrayList extension since it wasn't really necessary. I was only using it before to iterate with a for-each loop, but I could've done that with a regular for loop. I changed all declarations of InventoryArrayList with parametrized angle brackets to its regular call. I also completely eliminated the HashMap, it did not serve a purpose that couldn't be fulfilled by some unique method calls on the ArrayListInventory list. Now, removeButton works!